pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "book-translator"
version = "0.1.0"
description = "A tool to translate book pages from images to bilingual PDFs"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Utilities",
]
dependencies = [
    "Pillow>=9.0.0",
    "pytesseract>=0.3.10",
    "deep-translator>=1.10.0",
    "reportlab>=3.6.0",
    "opencv-python>=4.5.0",
]

[project.urls]
Homepage = "https://github.com/jarvasquezromo/book_translator"
Repository = "https://github.com/jarvasquezromo/book_translator"
Issues = "https://github.com/jarvasquezromo/book_translator/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[project.scripts]
book-translator = "src.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"